LayoutTests/inspector/debugger/debugger-autocontinue-on-syntax-error-expected.txt

CONSOLE MESSAGE: line 4: SyntaxError: Unexpected token ')'
CONSOLE MESSAGE: line 8: Iframe loaded
Tests that debugger won't stop on syntax errors even if "pause on uncaught exceptions" is on.

Debugger was enabled.
Debugger was disabled.
